+++ packages/nimble-components/src/icon-base/index.ts
@@ -1,0 +1,11 @@
+import { FoundationElement } from '@microsoft/fast-foundation';
+import type { NimbleIcon } from '@ni/nimble-tokens/dist-icons-esm/nimble-icons-inline';
+
+/**
+ * The base class for icon components
+ */
+export abstract class Icon extends FoundationElement {
+    public constructor(public icon: NimbleIcon) {
+        super();
+    }
+}